requirements.txt
Django==5.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
frappe-client==0.0.3
colorama==0.4.6
gunicorn==21.2.0
